Refactoring stage1 to facilitate  multiple implementations.
