+++ src/jsonparser.cpp
@@ -7,50 +7,56 @@
 #endif
 #include "simdjson/simdjson.h"
 
-// parse a document found in buf, need to preallocate ParsedJson.
-WARN_UNUSED
-int json_parse(const uint8_t *buf, size_t len, ParsedJson &pj, bool reallocifneeded) {
-  if (pj.bytecapacity < len) {
-    return simdjson::CAPACITY;
-  }
-  bool reallocated = false;
-  if(reallocifneeded) {
-#ifdef ALLOW_SAME_PAGE_BUFFER_OVERRUN
-	  // realloc is needed if the end of the memory crosses a page
-#ifdef _MSC_VER
-	  SYSTEM_INFO sysInfo; 
-	  GetSystemInfo(&sysInfo); 
-	  long pagesize = sysInfo.dwPageSize;
+
+// Responsible to select the best json_parse implementation
+int json_parse_dispatch(const uint8_t *buf, size_t len, ParsedJson &pj, bool reallocifneeded) {
+  // Versions for each implementation
+#ifdef __AVX2__
+  json_parse_functype* avx_implementation = &json_parse_implementation<simdjson::instruction_set::avx2>;
+#endif
+#ifdef __SSE4_2__
+  // json_parse_functype* sse4_2_implementation = &json_parse_implementation<simdjson::instruction_set::sse4_2>; // not implemented yet
+#endif
+#ifdef __ARM_NEON
+  json_parse_functype* neon_implementation = &json_parse_implementation<simdjson::instruction_set::neon>;
+#endif
+
+  // Determining which implementation is the more suitable
+  // Should be done at runtime. Does not make any sense on preprocessor.
+#ifdef __AVX2__
+  simdjson::instruction_set best_implementation = simdjson::instruction_set::avx2;
+#elif defined (__SSE4_2__)
+  simdjson::instruction_set best_implementation = simdjson::instruction_set::sse4_2;
+#elif defined (__ARM_NEON)
+  simdjson::instruction_set best_implementation = simdjson::instruction_set::neon;
 #else
-    long pagesize = sysconf (_SC_PAGESIZE); 
+  simdjson::instruction_set best_implementation = simdjson::instruction_set::none;
 #endif
-  //////////////
-  // We want to check that buf + len - 1 and buf + len - 1 + SIMDJSON_PADDING
-  // are in the same page.
-  // That is, we want to check that  
-  // (buf + len - 1) / pagesize == (buf + len - 1 + SIMDJSON_PADDING) / pagesize
-  // That's true if (buf + len - 1) % pagesize + SIMDJSON_PADDING < pagesize.
-  ///////////
-	 if ( (reinterpret_cast<uintptr_t>(buf + len - 1) % pagesize ) + SIMDJSON_PADDING < static_cast<uintptr_t>(pagesize) ) {
-#else // SIMDJSON_SAFE_SAME_PAGE_READ_OVERRUN
-     if(true) { // if not SIMDJSON_SAFE_SAME_PAGE_READ_OVERRUN, we always reallocate
+  
+  // Selecting the best implementation
+  switch (best_implementation) {
+#ifdef __AVX2__
+  case simdjson::instruction_set::avx2 :
+    json_parse_ptr = avx_implementation;
+    break;
+#elif defined (__SSE4_2__)
+  /*case simdjson::instruction_set::sse4_2 :
+    json_parse_ptr = sse4_2_implementation;
+    break;*/
+#elif defined (__ARM_NEON)
+  case simdjson::instruction_set::neon :
+    json_parse_ptr = neon_implementation;
+    break;
 #endif
-	     const uint8_t *tmpbuf  = buf;
-       buf = (uint8_t *) allocate_padded_buffer(len);
-       if(buf == NULL) return simdjson::MEMALLOC;
-       memcpy((void*)buf,tmpbuf,len);
-       reallocated = true;
-     }
+  default :
+    std::cerr << "No implemented simd instruction set supported" << std::endl;
+    return simdjson::UNEXPECTED_ERROR;
   }
-  int stage1_is_ok = find_structural_bits(buf, len, pj);
-  if(stage1_is_ok != simdjson::SUCCESS) {
-    pj.errorcode = stage1_is_ok;
-    return pj.errorcode;
-  } 
-  int res = unified_machine(buf, len, pj);
-  if(reallocated) { aligned_free((void*)buf);}
-  return res;
+
+  return json_parse_ptr(buf, len, pj, reallocifneeded);
 }
+
+json_parse_functype *json_parse_ptr = &json_parse_dispatch;
 
 WARN_UNUSED
 ParsedJson build_parsed_json(const uint8_t *buf, size_t len, bool reallocifneeded) {
